# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChainLance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List transactions
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /wallet/topup:
    post:
      operationId: requestTopup
      tags: [wallet]
      summary: Request ChainCoin top-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "201":
          description: Topup request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupRequest"
  /wallet/topup/list:
    get:
      operationId: listTopupRequests
      tags: [wallet]
      summary: List top-up requests
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Topup requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopupRequest"
  /wallet/topup/{id}/approve:
    patch:
      operationId: approveTopup
      tags: [wallet]
      summary: Approve/reject top-up request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupApprovalInput"
      responses:
        "200":
          description: Updated topup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupRequest"
  /wallet/withdraw:
    post:
      operationId: requestWithdraw
      tags: [wallet]
      summary: Request withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "201":
          description: Withdraw request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/transfer:
    post:
      operationId: transferCoins
      tags: [wallet]
      summary: Transfer ChainCoin to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferInput"
      responses:
        "200":
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # MESSAGES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        walletAddress: { type: ["string", "null"] }
        isSuspended: { type: boolean }
        createdAt: { type: string }
        chainCoinBalance: { type: number }
        level: { type: ["integer", "null"] }
        trustScore: { type: ["number", "null"] }
...
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        amount: { type: number }
        status: { type: string }
        txHash: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        createdAt: { type: string }
    TransactionList:
      type: object
      required: [transactions, total]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total: { type: integer }
    TopupInput:
      type: object
      required: [amount, txHash, paymentMethod]
      properties:
        amount: { type: number }
        txHash: { type: string }
        paymentMethod: { type: string }
        network: { type: string }
    TopupRequest:
      type: object
      required: [id, userId, amount, txHash, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        user:
          $ref: "#/components/schemas/User"
        amount: { type: number }
        txHash: { type: string }
        paymentMethod: { type: string }
        network: { type: ["string", "null"] }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
    TopupApprovalInput:
      type: object
      required: [status]
      properties:
        status: { type: string }
        adminNote: { type: string }
    WithdrawInput:
      type: object
      required: [amount, walletAddress, network]
      properties:
        amount: { type: number }
        walletAddress: { type: string }
        network: { type: string }
    TransferInput:
      type: object
      required: [toUserId, amount]
      properties:
        toUserId: { type: integer }
        amount: { type: number }
        note: { type: string }

    # MESSAGES
...